pyproject.toml
[tool.poetry]
name = "chatwithmycv"
version = "0.1.0"
description = "Build a simple chatbot which answers questions about your CV"
authors = ["Franzi <92818644+franziskavonalbedyll@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.28.1"
langchain = "^0.0.308"
python-dotenv = "^1.0.0"
chromadb = "^0.4.13"
tiktoken = "^0.5.1"
streamlit = "^1.27.2"
streamlit-chat = "^0.1.1"
azure-storage-blob = "^12.18.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
